requirements.txt
# 照片水印应用依赖项
# 基础依赖
PyQt6>=6.4.0
Pillow>=9.5.0
